Removing bunch of deprecated and unused stuff
