--- resteasy-jaxrs/src/main/java/org/jboss/resteasy/annotations/Suspend.java
@@ -1,22 +1,0 @@
-package org.jboss.resteasy.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
- * @version $Revision: 1 $
- * 
- * @deprecated Replaced by javax.ws.rs.container.Suspended
- * 
- * @see javax.ws.rs.container.Suspended
- */
-@Deprecated
-@Target({ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Suspend
-{
-   long value() default -1;
-}
