+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/core/InjectorFactoryImpl.java
@@ -2,7 +2,6 @@ package org.jboss.resteasy.core;
 
 import org.jboss.resteasy.annotations.Form;
 import org.jboss.resteasy.annotations.Query;
-import org.jboss.resteasy.annotations.Suspend;
 import org.jboss.resteasy.spi.ConstructorInjector;
 import org.jboss.resteasy.spi.InjectorFactory;
 import org.jboss.resteasy.spi.MethodInjector;
@@ -115,8 +114,6 @@ public class InjectorFactoryImpl implements InjectorFactory
             return new FormInjector(parameter.getType(), providerFactory);
          case MATRIX_PARAM:
             return new MatrixParamInjector(parameter.getType(), parameter.getGenericType(), parameter.getAccessibleObject(), parameter.getParamName(), parameter.getDefaultValue(), parameter.isEncoded(), parameter.getAnnotations(), providerFactory);
-         case SUSPEND:
-            return new SuspendInjector(parameter.getSuspendTimeout(), parameter.getType());
          case CONTEXT:
             return new ContextParameterInjector(null, parameter.getType(), parameter.getGenericType(), parameter.getAnnotations(), providerFactory);
          case SUSPENDED:
@@ -152,7 +149,6 @@ public class InjectorFactoryImpl implements InjectorFactory
       CookieParam cookie;
       FormParam formParam;
       Form form;
-      Suspend suspend;
       Suspended suspended;
       Query query;
 
@@ -232,10 +228,6 @@ public class InjectorFactoryImpl implements InjectorFactory
       else if (findAnnotation(annotations, org.jboss.resteasy.annotations.jaxrs.MatrixParam.class) != null)
       {
          return new MatrixParamInjector(type, genericType, injectTarget, defaultName, defaultVal, encode, annotations, providerFactory);
-      }
-      else if ((suspend = findAnnotation(annotations, Suspend.class)) != null)
-      {
-         return new SuspendInjector(suspend.value(), type);
       }
       else if (findAnnotation(annotations, Context.class) != null)
       {
