+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/plugins/delegates/MediaTypeHeaderDelegate.java
@@ -20,55 +20,6 @@ public class MediaTypeHeaderDelegate implements RuntimeDelegate.HeaderDelegate
       return parse(type);
    }
 
-   /*
-   public static MediaType parse(String type)
-   {
-      String params = null;
-      int idx = type.indexOf(";");
-      if (idx > -1)
-      {
-         params = type.substring(idx + 1).trim();
-         type = type.substring(0, idx);
-      }
-      String major = null;
-      String subtype = null;
-      String[] paths = type.split("/");
-      if (paths.length < 2 && type.equals("*"))
-      {
-         major = "*";
-         subtype = "*";
-
-      }
-      else if (paths.length != 2
-            || "".equals(paths[0]) || "".equals(paths[1])
-            || paths[0].contains(" ") || paths[1].contains(" "))
-      {
-         throw new IllegalArgumentException("Failure parsing MediaType string: " + type);
-      }
-      else if (paths.length == 2)
-      {
-         major = paths[0];
-         subtype = paths[1];
-      }
-      if (params != null && !params.equals(""))
-      {
-         HashMap<String, String> typeParams = new HashMap<String, String>();
-
-         int start = 0;
-
-         while (start < params.length())
-         {
-            start = HeaderParameterParser.setParam(typeParams, params, start);
-         }
-         return new MediaType(major, subtype, typeParams);
-      }
-      else
-      {
-         return new MediaType(major, subtype);
-      }
-   }
-   */
-
    protected static boolean isValid(String str)
    {
       if (str == null || str.length() == 0) return false;
