+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/plugins/interceptors/ClientContentEncodingAnnotationFeature.java
@@ -20,6 +20,7 @@ public class ClientContentEncodingAnnotationFeature implements DynamicFeature
    @Override
    public void configure(ResourceInfo resourceInfo, FeatureContext configurable)
    {
+      @SuppressWarnings("rawtypes")
       final Class declaring = resourceInfo.getResourceClass();
       final Method method = resourceInfo.getResourceMethod();
