+++ resteasy-jaxrs/src/main/java/org/jboss/resteasy/plugins/interceptors/RoleBasedSecurityFeature.java
@@ -18,6 +18,7 @@ public class RoleBasedSecurityFeature implements DynamicFeature
    @Override
    public void configure(ResourceInfo resourceInfo, FeatureContext configurable)
    {
+      @SuppressWarnings("rawtypes")
       final Class declaring = resourceInfo.getResourceClass();
       final Method method = resourceInfo.getResourceMethod();
