--- resteasy-jaxrs/src/main/java/org/jboss/resteasy/util/IsAssignableFrom.java
@@ -1,9 +1,0 @@
-package org.jboss.resteasy.util;
-
-/**
- * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
- * @version $Revision: 1 $
- */
-public class IsAssignableFrom
-{
-}
